refactoring rst2markdown
