+++ converter/rst/heading.py
@@ -1,0 +1,34 @@
+import re
+
+
+class Heading(object):
+    def __init__(self, source_string, caret_token):
+        self.str = source_string
+        self._caret_token = caret_token
+        self._heading1_re = re.compile(r"""^(?P<content>.*?\n)?(?:=)+\s*$""", flags=re.MULTILINE)
+        self._heading2_re = re.compile(r"""^(?P<content>.*?\n)?(?:-)+\s*$""", flags=re.MULTILINE)
+        self._heading3_re = re.compile(r"""^(?P<content>.*?\n)?(?:~)+\s*$""", flags=re.MULTILINE)
+        self._heading4_re = re.compile(r"""^(?P<content>.*?\n)?(?:")+\s*$""", flags=re.MULTILINE)
+
+    def _heading1(self, matchobj):
+        return ''
+
+    def _heading2(self, matchobj):
+        content = matchobj.group('content')
+        return f'{self._caret_token}## {content}'
+
+    def _heading3(self, matchobj):
+        content = matchobj.group('content')
+        return f'{self._caret_token}{self._caret_token}### {content}'
+
+    def _heading4(self, matchobj):
+        content = matchobj.group('content')
+        return f'{self._caret_token}#### {content}'
+
+    def convert(self):
+        output = self.str
+        output = self._heading1_re.sub(self._heading1, output)
+        output = self._heading2_re.sub(self._heading2, output)
+        output = self._heading3_re.sub(self._heading3, output)
+        output = self._heading4_re.sub(self._heading4, output)
+        return output
