+++ converter/rst2markdown.py
@@ -3,8 +3,6 @@ import pathlib
 import re
 import uuid
 
-from converter.guides.tools import slugify
-from string import Template
 from collections import namedtuple
 
 from converter.rst.table import Table
@@ -12,59 +10,13 @@ from converter.rst.image import Image
 from converter.rst.sidebar import Sidebar
 from converter.rst.external_link import ExternalLink
 from converter.rst.footnote import Footnote
+from converter.rst.heading import Heading
+from converter.rst.definition import Definition
+from converter.rst.inlineav import InlineAv
+from converter.rst.avembed import AvEmbed
 
 AssessmentData = namedtuple('AssessmentData', ['id', 'name', 'type', 'points', 'ex_data'])
-IframeImage = namedtuple('IframeImage', ['src', 'path', 'content'])
 OPEN_DSA_CDN = 'https://global.codio.com/opendsa/v3'
-GUIDES_CDN = '//static-assets.codio.com/guides/opendsa/v1'
-MATHJAX_CDN = '//cdnjs.cloudflare.com/ajax/libs/mathjax/2.7.1'
-JSAV_IFRAME_SUBPATH = 'jsav/iframe/v6/'
-JSAV_IMAGE_IFRAME = f"""
-<html>
-<head>
-<title>$title</title>
-<link rel="stylesheet" href="{GUIDES_CDN}/haiku.css" type="text/css" />
-<link rel="stylesheet" href="{GUIDES_CDN}/normalize.css" type="text/css" />
-<link rel="stylesheet" href="{GUIDES_CDN}/JSAV.css" type="text/css" />
-<link rel="stylesheet" href="{GUIDES_CDN}/odsaMOD-min.css" type="text/css" />
-<link rel="stylesheet" href="{GUIDES_CDN}/jquery-ui.css" type="text/css" />
-<link rel="stylesheet" href="{GUIDES_CDN}/odsaStyle-min.css" type="text/css" />
-
-<script type="text/javascript" src="{GUIDES_CDN}/jquery-2.1.4.min.js"></script>
-<script type="text/javascript" src="{MATHJAX_CDN}/MathJax.js?config=TeX-AMS-MML_HTMLorMML"></script>
-<script type="text/javascript" src="{GUIDES_CDN}/jquery-ui.min.js"></script>
-<script type="text/javascript" src="{GUIDES_CDN}/jquery.transit.js"></script>
-<script type="text/javascript" src="{GUIDES_CDN}/raphael.js"></script>
-<script type="text/javascript" src="{GUIDES_CDN}/JSAV-min.js"></script>
-<script type="text/javascript" src="{GUIDES_CDN}/odsaUtils-min.js"></script>
-<script type="text/javascript" src="{GUIDES_CDN}/odsaMOD-min.js"></script>
-<script type="text/javascript" src="{GUIDES_CDN}/d3.min.js"></script>
-<script type="text/javascript" src="{GUIDES_CDN}/d3-selection-multi.v1.min.js"></script>
-<script type="text/javascript" src="{GUIDES_CDN}/dataStructures.js"></script>
-<script type="text/javascript" src="{GUIDES_CDN}/conceptMap.js"></script>
-</head>
-<body>
-$content
-<script>
-window.addEventListener("load", sendPostMessage);
-var element = document.getElementById('$name');
-if (element) {{
-    element.addEventListener("resize", sendPostMessage);
-    element.addEventListener("click", sendPostMessage);
-}}
-var height;
-function sendPostMessage() {{
-    if (height !== element.offsetHeight && element) {{
-        height = element.offsetHeight;
-        window.parent.postMessage(
-            JSON.stringify({{frameHeight: height, frameId: "$name", status: 'iframe', av: "$name"}}), '*'
-        );
-    }}
-}}
-</script>
-</body>
-</html>
-"""
 
 
 class Rst2Markdown(object):
@@ -78,10 +30,6 @@ class Rst2Markdown(object):
         self.lines_array = lines_array
         self._exercises = exercises
         self.workspace_dir = workspace_dir
-        self._heading1_re = re.compile(r"""^(?P<content>.*?\n)?(?:=)+\s*$""", flags=re.MULTILINE)
-        self._heading2_re = re.compile(r"""^(?P<content>.*?\n)?(?:-)+\s*$""", flags=re.MULTILINE)
-        self._heading3_re = re.compile(r"""^(?P<content>.*?\n)?(?:~)+\s*$""", flags=re.MULTILINE)
-        self._heading4_re = re.compile(r"""^(?P<content>.*?\n)?(?:")+\s*$""", flags=re.MULTILINE)
         self._list_re = re.compile(r"""^( *)(?P<type>[*+\-]|[0-9#]+[\.]) [^\n]*(?:\n(?!\1\2|\S)[^\n]*)*""",
                                    flags=re.MULTILINE)
         self._ext_links_re = re.compile(r"""`(?P<name>.*?)\n?<(?P<ref>https?:.*?)>`_""")
@@ -99,37 +47,11 @@ class Rst2Markdown(object):
         self._tip_re = re.compile(
             r"""^\.{2} tip:{2} *\n(?P<content>(?:\n* +[^\n]+\n*)*)""", flags=re.MULTILINE)
         self._epigraph_re = re.compile(r"""[ ]*\.\. epigraph:: *\n*(?P<content>(?: +[^\n]+\n*)*)""")
-        self._inlineav_re = re.compile(
-            r"""(\.\. _.*?:\n^$\n)?\.\. inlineav:: (?P<name>.*?) (?P<type>.*?)(?P<options>:.*?: .*?\n)+(?=\S|$)""",
-            flags=re.MULTILINE + re.DOTALL
-        )
-        self._avembed_re = re.compile(
-            r"""\s*\.\. avembed:: (?P<name>.*?) (?P<type>[a-z]{2})\n(?P<options>(\s+:.*?:\s+.*\n)+)?"""
-        )
         self._code_include_re = re.compile(r"""\.\. codeinclude:: (?P<path>.*?)\n(?P<options>(?: +:.*?: \S*\n)+)?""")
         self._extrtoolembed_re = re.compile(
             r"""^$\n^.*?\n-+\n\n?\.\. extrtoolembed:: '(?P<name>.*?)'\n( *:.*?: .*?\n)?(?=\S|$)""", flags=re.MULTILINE)
         self._term_def_re = re.compile(r"""^:(?P<term>[^:\n]+): *\n(?P<content>(?: +[^\n]+\n*)*)""", flags=re.MULTILINE)
         self._lineblock_re = re.compile(r"""^((?: {2,})?\|)[^\n]*(?:\n(?:\1| {2,})[^\n]+)*""", flags=re.MULTILINE)
-        self._definition_re = re.compile(
-            r"""^(?!\s|\d\. |#\. |\* |- |\.\. ):?(?P<term>[^\n]+)\n(?P<content> {2,}[^\n]+(?:\n {2,}[^\n]+\s*)*\n+)""",
-            flags=re.MULTILINE
-        )
-
-    def _heading1(self, matchobj):
-        return ''
-
-    def _heading2(self, matchobj):
-        content = matchobj.group('content')
-        return f'{self._caret_token}## {content}'
-
-    def _heading3(self, matchobj):
-        content = matchobj.group('content')
-        return f'{self._caret_token}{self._caret_token}### {content}'
-
-    def _heading4(self, matchobj):
-        content = matchobj.group('content')
-        return f'{self._caret_token}#### {content}'
 
     def _list(self, matchobj):
         caret_token = self._caret_token
@@ -226,24 +148,6 @@ class Rst2Markdown(object):
         content = '\n'.join(out)
         return f'<div style="padding: 10px 30px;">{caret_token}{content}{caret_token}</div>{caret_token}{caret_token}'
 
-    def _definition(self, matchobj):
-        caret_token = self._caret_token
-        term = matchobj.group('term')
-        term = term.strip('**').strip()
-        content = matchobj.group('content')
-
-        space = re.search('\n *', content)
-        space_count = len(space.group(0)) - 1
-        space_regex = f"\n^ {{{space_count}}}"
-        content = re.sub(space_regex, '', content, flags=re.MULTILINE)
-
-        lines = content.split('\n')
-        out = []
-        for line in lines:
-            out.append(f'{line.strip()}{caret_token}')
-        content = '\n'.join(out)
-        return f'{caret_token}{caret_token}**{term}**{caret_token}{content}{caret_token}{caret_token}'
-
     def _lineblock(self, matchobj):
         caret_token = self._caret_token
         content = matchobj.group(0)
@@ -284,109 +188,6 @@ class Rst2Markdown(object):
         assessment = AssessmentData(assessment_id, name, 'test', 20, ex_data)
         self._assessments.append(assessment)
         return ''
-
-    def _avembed(self, matchobj):
-        caret_token = self._caret_token
-        file_name = matchobj.group('name')
-        # todo: check different type
-        av_type = matchobj.group('type')
-        # todo: verify options
-        # options = matchobj.group('options')
-        name = pathlib.Path(file_name).stem
-
-        # todo: future todos: upload and use cdn path
-
-        assessment_id = f'custom-{name.lower()}'
-        assessment = AssessmentData(assessment_id, name, 'custom', 1, {'question': 'Resolve the challenge above'})
-        self._assessments.append(assessment)
-
-        return f'{caret_token}<iframe id="{name}_iframe" src="{OPEN_DSA_CDN}/{file_name}' \
-               f'?selfLoggingEnabled=false&localMode=true&JXOP-debug=true&JOP-lang=en&JXOP-code=java' \
-               f'&scoringServerEnabled=false&threshold=5&amp;points=1.0&required=True" ' \
-               f'class="embeddedExercise" width="950" height="800" data-showhide="show" scrolling="yes" ' \
-               f'style="position: relative; top: 0px;">Your browser does not support iframes.</iframe>' \
-               f'{caret_token}{caret_token}{{Check It!|assessment}}({assessment_id}){caret_token}{caret_token}'
-
-    def _inlineav(self, matchobj):
-        images = {}
-        caption = ""
-        caret_token = self._caret_token
-        option_re = re.compile(':([^:]+): (.+)')
-        name = matchobj.group('name')
-        av_type = matchobj.group('type')
-        av_type = av_type.strip()
-        options = matchobj.group('options')
-        options = options.split('\n')
-        for line in options:
-            line = line.strip()
-            opt_match = option_re.match(line)
-            if opt_match:
-                images[opt_match[1]] = opt_match[2]
-            elif line.strip():
-                caption += f'{line} '
-        script_opt = images.get('scripts', '')
-        script_opt = script_opt.split()
-        css_opt = images.get('links', '')
-        css_opt = css_opt.split()
-        self._figure_counter += 1
-        counter = f'{self._chapter_num}.{self._subsection_num}.{self._figure_counter}'
-        if caption:
-            caption = caption.strip()
-            caption = f'<center>{counter} {caption}</center><br/>{caret_token}{caret_token}'
-        scripts = ''.join(list(map(lambda x: f'<script type="text/javascript" src="{OPEN_DSA_CDN}/{x}">'
-                                             f'</script>{caret_token}', script_opt)))
-        css_links = ''.join(list(map(lambda x: f'<link rel="stylesheet" type="text/css" href="{OPEN_DSA_CDN}/{x}"/>'
-                                               f'{caret_token}', css_opt)))
-        iframe_name = slugify(name)
-        """
-        reason for subpath - some dynamic images have relative imports like ../../../SourceCode/target_file
-        and it allow load it in correct way from cdn root
-        """
-
-        iframe_src = f'{OPEN_DSA_CDN}/{JSAV_IFRAME_SUBPATH}{iframe_name}.html'
-        iframe_content = ''
-
-        if av_type == 'dgm':
-            iframe_content = f'{css_links}\n' \
-                             f'<div style="margin: 0" id="{name}"></div>\n' \
-                             f'{scripts}\n'
-        if av_type == 'ss':
-            iframe_content = f'{css_links}\n' \
-                             f'<div style="margin: 0" id="{name}" class="ssAV">\n' \
-                             f'<span class="jsavcounter"></span>\n' \
-                             f'<a class="jsavsettings" href="#">Settings</a>\n' \
-                             f'<div class="jsavcontrols"></div>\n' \
-                             f'<p class="jsavoutput jsavline"></p>\n' \
-                             f'<div class="jsavcanvas"></div>\n' \
-                             f'</div>\n{scripts}\n'
-
-        iframe_content = re.sub(caret_token, '\n', iframe_content)
-        iframe_body = Template(JSAV_IMAGE_IFRAME).substitute(title=name, content=iframe_content, name=name)
-
-        self._iframe_images.append(IframeImage(iframe_src, f'{JSAV_IFRAME_SUBPATH}{iframe_name}.html', iframe_body))
-        iframe_height = self.detect_height_from_css(css_opt, name)
-
-        return f'{caret_token}<iframe id="{name}_iframe" src="{iframe_src}" ' \
-               f'width="900" height="{iframe_height}" scrolling="no" ' \
-               f'style="position: relative; top: 0px; border: 0; margin: 0; overflow: hidden;">' \
-               f'Your browser does not support iframes.</iframe>{caret_token}' \
-               f'<br/>{caret_token}{caption}{caret_token}'
-
-    def detect_height_from_css(self, css_names, image_id):
-        for css_name in css_names:
-            css_path = self.workspace_dir.joinpath(css_name)
-            if not css_path.exists():
-                continue
-            with open(css_path, 'r') as file:
-                css_content = file.read().replace('\n', '')
-                result = re.match(rf"""#{image_id}(?P<content>.*?)}}""", css_content)
-                if result:
-                    css_height_opt = result.group('content')
-                    result_height = re.match(r""".*{.*height(\s)*:(\s)*(?P<height>\d+)px""", css_height_opt)
-                    if result_height:
-                        iframe_height = int(result_height.group('height'))
-                        return iframe_height + 30
-        return 250
 
     def _code_include(self, matchobj):
         options = {}
@@ -470,18 +271,27 @@ class Rst2Markdown(object):
         output = re.sub(r"\+\+", "\\+\\+", output)
         output = re.sub(r"^\|$", "<br/>", output, flags=re.MULTILINE)
         output = self._extrtoolembed_re.sub(self._extrtoolembed, output)
-        output = self._heading1_re.sub(self._heading1, output)
-        output = self._heading2_re.sub(self._heading2, output)
-        output = self._heading3_re.sub(self._heading3, output)
-        output = self._heading4_re.sub(self._heading4, output)
-        output = self._definition_re.sub(self._definition, output)
+        output = Heading(output, self._caret_token).convert()
+        output = Definition(output, self._caret_token).convert()
         output = self._lineblock_re.sub(self._lineblock, output)
         output = self._topic_re.sub(self._topic, output)
         output = self._tip_re.sub(self._tip, output)
         output = Image(output, self._caret_token, self._chapter_num, self._subsection_num).convert()
         output = re.sub(r".. _[\S ]+:", "", output)  # todo: line moved below. removes tags for images
-        output = self._inlineav_re.sub(self._inlineav, output)
-        output = self._avembed_re.sub(self._avembed, output)
+        output, figure_counter, iframe_images = InlineAv(output, self._caret_token,
+                                                         self.workspace_dir, self._chapter_num,
+                                                         self._subsection_num, OPEN_DSA_CDN,
+                                                         self._figure_counter
+                                                         ).convert()
+        if figure_counter:
+            self._figure_counter = figure_counter
+
+        if iframe_images:
+            self._iframe_images.append(iframe_images)
+
+        output, assessments = AvEmbed(output, self._caret_token, OPEN_DSA_CDN).convert()
+        # if assessments:
+        #     self._assessments.append(assessments)
         output = self._list_re.sub(self._list, output)
         output = self._ext_links_re.sub(self._ext_links, output)
         output = self._ref_re.sub(self._ref, output)
